+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reorder common tests, so instead of running `default apps` and then `basic`, now it first runs `cluster basic`, then `default apps`, and then `workload basic`.
 - Run all common tests before the upgrade in the CAPA upgrade suite.
 - Increase node pool minSize from 2 to 3 in the CAPA upgrade suite, as we have noticed that 2 nodes are not enough for all the Pods after teh upgrade, and cluster-autoscaler does not scale up the cluster quickly enough for the test to pass.
+- Run CAPA standard tests in order.
 
 ### Removed